More functional tests : - Delete user

+++ front_api/rakuten_api.py
@@ -313,20 +313,28 @@ async def delete_user(user: str, admin: User = Depends(get_admin)):
 
     engine = get_engine()
     with engine.connect() as connection:
-        statement = text('DELETE FROM users WHERE users.user = :user')
-        try:
-            connection.execute(statement,{"user": user})
-            # Default rollback behavior with alchemy, so commit change !!!!
-            connection.commit()
-        except ValueError:
-            connection.rollback()
+
+        select_statement = text('SELECT * FROM users WHERE users.user = :user')
+        result = connection.execute(select_statement, {"user": user})
+        user_exists = result.fetchone() is not None
+        if user_exists:
+            statement = text('DELETE FROM users WHERE users.user = :user')
+            try:
+                connection.execute(statement,{"user": user})
+                # Default rollback behavior with alchemy, so commit change !!!!
+                connection.commit()
+            except ValueError:
+                connection.rollback()
+                raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="SQL Error"
+            )
+        else:
             raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Erreur SQL"
-        )
-        # A METTRE ERREUR SQL PERSONNALISEE
-    # return {'DELETE FROM users WHERE users.user = ''{user}'''.format(user=user)}
-    return {"detail": "User successfully deleted"}
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail="User {} not found".format(user)
+            )
+    return {"detail": "User {} successfully deleted".format(user)}
 
 
 @server.get('/users/{user_id:int}', response_model=InputUser, tags=['users'])
